+++ crates/interpreter/src/instructions/host.rs
@@ -1,4 +1,5 @@
 use crate::primitives::{Bytes, Spec, SpecId::*, B160, B256, U256};
+use crate::MAX_INITCODE_SIZE;
 use crate::{
     alloc::vec::Vec,
     gas::{self, COLD_ACCOUNT_ACCESS_COST, WARM_STORAGE_READ_COST},
@@ -92,7 +93,7 @@ pub fn extcodecopy<SPEC: Spec>(interpreter: &mut Interpreter, host: &mut dyn Hos
     }
     let (code, is_cold) = ret.unwrap();
 
-    let len = as_usize_or_fail!(interpreter, len_u256, InstructionResult::OutOfGas);
+    let len = as_usize_or_fail!(interpreter, len_u256, InstructionResult::InvalidOperandOOG);
     gas_or_fail!(
         interpreter,
         gas::extcodecopy_cost::<SPEC>(len as u64, is_cold)
@@ -100,7 +101,11 @@ pub fn extcodecopy<SPEC: Spec>(interpreter: &mut Interpreter, host: &mut dyn Hos
     if len == 0 {
         return;
     }
-    let memory_offset = as_usize_or_fail!(interpreter, memory_offset, InstructionResult::OutOfGas);
+    let memory_offset = as_usize_or_fail!(
+        interpreter,
+        memory_offset,
+        InstructionResult::InvalidOperandOOG
+    );
     let code_offset = min(as_usize_saturated!(code_offset), code.len());
     memory_resize!(interpreter, memory_offset, len);
 
@@ -144,7 +149,7 @@ pub fn sload<SPEC: Spec>(interpreter: &mut Interpreter, host: &mut dyn Host) {
 }
 
 pub fn sstore<SPEC: Spec>(interpreter: &mut Interpreter, host: &mut dyn Host) {
-    check!(interpreter, !interpreter.is_static);
+    check_staticcall!(interpreter);
 
     pop!(interpreter, index, value);
     let ret = host.sstore(interpreter.contract.address, index, value);
@@ -163,16 +168,16 @@ pub fn sstore<SPEC: Spec>(interpreter: &mut Interpreter, host: &mut dyn Host) {
     }
 }
 
-pub fn log<const N: u8, SPEC: Spec>(interpreter: &mut Interpreter, host: &mut dyn Host) {
-    check!(interpreter, !interpreter.is_static);
+pub fn log<const N: u8>(interpreter: &mut Interpreter, host: &mut dyn Host) {
+    check_staticcall!(interpreter);
 
     pop!(interpreter, offset, len);
-    let len = as_usize_or_fail!(interpreter, len, InstructionResult::OutOfGas);
+    let len = as_usize_or_fail!(interpreter, len, InstructionResult::InvalidOperandOOG);
     gas_or_fail!(interpreter, gas::log_cost(N, len as u64));
     let data = if len == 0 {
         Bytes::new()
     } else {
-        let offset = as_usize_or_fail!(interpreter, offset, InstructionResult::OutOfGas);
+        let offset = as_usize_or_fail!(interpreter, offset, InstructionResult::InvalidOperandOOG);
         memory_resize!(interpreter, offset, len);
         Bytes::copy_from_slice(interpreter.memory.get_slice(offset, len))
     };
@@ -194,7 +199,7 @@ pub fn log<const N: u8, SPEC: Spec>(interpreter: &mut Interpreter, host: &mut dy
 }
 
 pub fn selfdestruct<SPEC: Spec>(interpreter: &mut Interpreter, host: &mut dyn Host) {
-    check!(interpreter, !interpreter.is_static);
+    check_staticcall!(interpreter);
     pop_address!(interpreter, target);
 
     let res = host.selfdestruct(interpreter.contract.address, target);
@@ -217,7 +222,7 @@ pub fn create<const IS_CREATE2: bool, SPEC: Spec>(
     interpreter: &mut Interpreter,
     host: &mut dyn Host,
 ) {
-    check!(interpreter, !interpreter.is_static);
+    check_staticcall!(interpreter);
     if IS_CREATE2 {
         // EIP-1014: Skinny CREATE2
         check!(interpreter, SPEC::enabled(PETERSBURG));
@@ -226,12 +231,24 @@ pub fn create<const IS_CREATE2: bool, SPEC: Spec>(
     interpreter.return_data_buffer = Bytes::new();
 
     pop!(interpreter, value, code_offset, len);
-    let len = as_usize_or_fail!(interpreter, len, InstructionResult::OutOfGas);
+    let len = as_usize_or_fail!(interpreter, len, InstructionResult::InvalidOperandOOG);
 
     let code = if len == 0 {
         Bytes::new()
     } else {
-        let code_offset = as_usize_or_fail!(interpreter, code_offset, InstructionResult::OutOfGas);
+        let code_offset = as_usize_or_fail!(
+            interpreter,
+            code_offset,
+            InstructionResult::InvalidOperandOOG
+        );
+        // EIP-3860: Limit and meter initcode
+        if SPEC::enabled(SHANGHAI) {
+            if len > MAX_INITCODE_SIZE {
+                interpreter.instruction_result = InstructionResult::CreateInitcodeSizeLimit;
+                return;
+            }
+            gas!(interpreter, gas::initcode_cost(len as u64));
+        }
         memory_resize!(interpreter, code_offset, len);
         Bytes::copy_from_slice(interpreter.memory.get_slice(code_offset, len))
     };
@@ -347,18 +364,23 @@ pub fn call_inner<SPEC: Spec>(
 
     pop!(interpreter, in_offset, in_len, out_offset, out_len);
 
-    let in_len = as_usize_or_fail!(interpreter, in_len, InstructionResult::OutOfGas);
+    let in_len = as_usize_or_fail!(interpreter, in_len, InstructionResult::InvalidOperandOOG);
     let input = if in_len != 0 {
-        let in_offset = as_usize_or_fail!(interpreter, in_offset, InstructionResult::OutOfGas);
+        let in_offset =
+            as_usize_or_fail!(interpreter, in_offset, InstructionResult::InvalidOperandOOG);
         memory_resize!(interpreter, in_offset, in_len);
         Bytes::copy_from_slice(interpreter.memory.get_slice(in_offset, in_len))
     } else {
         Bytes::new()
     };
 
-    let out_len = as_usize_or_fail!(interpreter, out_len, InstructionResult::OutOfGas);
+    let out_len = as_usize_or_fail!(interpreter, out_len, InstructionResult::InvalidOperandOOG);
     let out_offset = if out_len != 0 {
-        let out_offset = as_usize_or_fail!(interpreter, out_offset, InstructionResult::OutOfGas);
+        let out_offset = as_usize_or_fail!(
+            interpreter,
+            out_offset,
+            InstructionResult::InvalidOperandOOG
+        );
         memory_resize!(interpreter, out_offset, out_len);
         out_offset
     } else {