+++ crates/primitives/src/db/components/block_hash.rs
@@ -2,7 +2,9 @@
 //! it is used inside [crate::db::DatabaseComponents`]
 
 use crate::{B256, U256};
+use alloc::sync::Arc;
 use auto_impl::auto_impl;
+use core::ops::Deref;
 
 #[auto_impl(& mut, Box)]
 pub trait BlockHash {
@@ -28,5 +30,16 @@ where
 
     fn block_hash(&mut self, number: U256) -> Result<B256, Self::Error> {
         BlockHashRef::block_hash(*self, number)
+    }
+}
+
+impl<T> BlockHash for Arc<T>
+where
+    T: BlockHashRef,
+{
+    type Error = <T as BlockHashRef>::Error;
+
+    fn block_hash(&mut self, number: U256) -> Result<B256, Self::Error> {
+        self.deref().block_hash(number)
     }
 }